Simplified EERs to take optional sourceMention
